Move some login module files
