+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/data/source/api/mappers/SignIn.kt
@@ -1,0 +1,8 @@
+package com.gdavidpb.tuindice.login.data.source.api.mappers
+
+import com.gdavidpb.tuindice.login.domain.model.SignIn
+import com.gdavidpb.tuindice.login.data.source.api.responses.SignInResponse
+
+fun SignInResponse.toSignIn() = SignIn(
+	token = token
+)
