+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/data/source/api/responses/SignInResponse.kt
@@ -1,4 +1,4 @@
-package com.gdavidpb.tuindice.login.data.api.responses
+package com.gdavidpb.tuindice.login.data.source.api.responses
 
 import com.google.gson.annotations.SerializedName
