+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/domain/usecase/SignInUseCase.kt
@@ -3,10 +3,10 @@ package com.gdavidpb.tuindice.login.domain.usecase
 import com.gdavidpb.tuindice.base.domain.repository.*
 import com.gdavidpb.tuindice.base.domain.usecase.base.EventUseCase
 import com.gdavidpb.tuindice.base.utils.ConfigKeys
+import com.gdavidpb.tuindice.base.utils.Topics
 import com.gdavidpb.tuindice.base.utils.annotations.Timeout
 import com.gdavidpb.tuindice.base.utils.extensions.*
 import com.gdavidpb.tuindice.login.domain.model.SignInRequest
-import com.gdavidpb.tuindice.base.utils.Topics
 import com.gdavidpb.tuindice.login.domain.repository.RemoteRepository
 import com.gdavidpb.tuindice.login.domain.usecase.error.SignInError
