+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/presentation/viewmodel/SignInViewModel.kt
@@ -1,16 +1,16 @@
 package com.gdavidpb.tuindice.login.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
+import com.gdavidpb.tuindice.base.domain.usecase.SignOutUseCase
 import com.gdavidpb.tuindice.base.domain.usecase.SyncUseCase
 import com.gdavidpb.tuindice.base.domain.usecase.error.SyncError
 import com.gdavidpb.tuindice.base.utils.extensions.LiveCompletable
 import com.gdavidpb.tuindice.base.utils.extensions.LiveEvent
 import com.gdavidpb.tuindice.base.utils.extensions.LiveResult
-import com.gdavidpb.tuindice.login.domain.model.SignInRequest
 import com.gdavidpb.tuindice.base.utils.extensions.execute
+import com.gdavidpb.tuindice.login.domain.model.SignInRequest
 import com.gdavidpb.tuindice.login.domain.usecase.ReSignInUseCase
 import com.gdavidpb.tuindice.login.domain.usecase.SignInUseCase
-import com.gdavidpb.tuindice.base.domain.usecase.SignOutUseCase
 import com.gdavidpb.tuindice.login.domain.usecase.error.SignInError
 
 class SignInViewModel(
