+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/presentation/viewmodel/SplashViewModel.kt
@@ -1,8 +1,8 @@
 package com.gdavidpb.tuindice.login.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
-import com.gdavidpb.tuindice.base.utils.extensions.LiveResult
 import com.gdavidpb.tuindice.base.domain.model.StartUpAction
+import com.gdavidpb.tuindice.base.utils.extensions.LiveResult
 import com.gdavidpb.tuindice.base.utils.extensions.execute
 import com.gdavidpb.tuindice.login.domain.usecase.StartUpUseCase
 import com.gdavidpb.tuindice.login.domain.usecase.error.StartUpError
